+++ source/core/coding/ht_block_decoding_avx2.cpp
@@ -573,7 +573,7 @@ void j2k_codeblock::dequantize(uint8_t ROIshift) const {
         }
 
         assert(pLSB >= 0);  // assure downshift is not negative
-        *dst = static_cast<int32_t>(*val);
+        *dst = static_cast<float>(*val);
         val++;
         dst++;
       }
@@ -650,7 +650,7 @@ void j2k_codeblock::dequantize(uint8_t ROIshift) const {
         if (sign) {
           *val = -(*val & INT32_MAX);
         }
-        *dst = static_cast<int32_t>(*val);
+        *dst = static_cast<float>(*val);
         val++;
         dst++;
       }
